fixed import bug and moved general utility methods into misc folder
