+++ client/src/tests/utils/misc.js
@@ -1,0 +1,2 @@
+
+export const generateRandomString = () => Math.random().toString(36).substr(2, 5);
